hal_interface2.0: fix settings.find methor access conflict
updateMetaParams & processResults may access inputParams.setting
at the same time, it causes crash when preview.

+++ hal_interface2.0/interface/rkisp1_aiq_types.h
@@ -71,7 +71,7 @@ extern "C" {
 
 typedef struct
 {
-  unsigned int coeff[9U];
+  float coeff[9U];
 } rk_aiq_3x3Matrix;
 
 typedef struct
